Start refactor toward a domain driven design folder structure
Created folder structure and mostly empty files. Made the first move of the user model into domain/users/model.py
